Se agrega pruebas unitarias punto 1
