+++ apiFestivos.test/apiFestivosTest.cs
@@ -85,5 +85,49 @@ namespace apiFestivos.test
             var fechaEsperada = new DateTime(1999, 3, 28);
             Assert.True(result.Any(f => f.Fecha == fechaEsperada), "festivo no encontrado");
         }
+        [Fact]
+        public async Task verificaCoincideFechaConFestivo()
+        {
+            // Arrange
+            var festivos = new List<Festivo>
+            {
+                new Festivo { Id = 1, Nombre = "Año Nuevo 1", Dia = 1, Mes = 1, IdTipo = 1 },
+                new Festivo { Id = 2, Nombre = "Santos Reyes", Dia = 6, Mes = 1, IdTipo = 2 },
+                new Festivo { Id = 3, Nombre = "Ascensión del Señor", DiasPascua = 40, IdTipo = 4 },
+                new Festivo { Id = 4, Nombre = "Año Nuevo 4", Dia = 4, Mes = 1, IdTipo = 2 },
+                new Festivo { Id = 5, Nombre = "Año Nuevo 5", Dia = 5, Mes = 1, IdTipo = 3 }
+
+            };
+            festivoRepositorioMock.Setup(repo => repo.ObtenerTodos()).ReturnsAsync(festivos);
+
+            var fechaFestiva = new DateTime(2024, 1, 1);
+
+            // Act
+            var esFestivo = await festivoServicio.EsFestivo(fechaFestiva);
+
+            // Assert
+            Assert.True(esFestivo, "La fecha debería ser festiva, pero no lo fue.");
+        }
+
+        [Fact]
+        public async Task FechaNoFestivaDeberiaDevolverFalse()
+        {
+            // Arrange
+            var festivos = new List<Festivo>
+            {
+                new Festivo { Id = 1, Nombre = "Año Nuevo", Dia = 1, Mes = 1, IdTipo = 1 },
+                new Festivo { Id = 2, Nombre = "Día de la Independencia", Dia = 4, Mes = 7, IdTipo = 1 }
+            };
+
+            festivoRepositorioMock.Setup(repo => repo.ObtenerTodos()).ReturnsAsync(festivos);
+
+            var fechaNoFestiva = new DateTime(2024, 2, 15); // Una fecha que no está en la lista.
+
+            // Act
+            var esFestivo = await festivoServicio.EsFestivo(fechaNoFestiva);
+
+            // Assert
+            Assert.False(esFestivo, "La fecha no debería ser festiva, pero el método devolvió true.");
+        }
     }
 }
